Rename GitHub action to publish packages
Renamed the GitHub action from 'Release' to 'Publish to GitHub Package Registry'. It now triggers on push events to any tags instead of on creation of a release. This change streamlines the process of publishing the package using the GitHub token and ensures that the package always gets published on the latest version.