Update trigger for GitHub workflows in publish.yml
The trigger condition for GitHub workflows has been changed in the publish.yml file from any new tags to the push to master branch. This adjustment helps to facilitate a workflow designed for continuous integration/delivery, where updated code is automatically prepared for a release whenever changes are made to the master branch.